Pre-Extraction helper, structs into single file
This commit adds a extraction helper script (extr_helper.sh) for easiliy pre-extract data from source file, i.e. convert XLS to HTML and extract data from PDFs as JSON. It also takes care of previous files and data.
Also moves all structs into a single file "structs.rb."
