--- loffice_version.txt
@@ -1,2 +1,0 @@
-LibreOffice 5.1.2.1.0 10m0(Build:1)
-
